% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flooding.R
\name{flooding}
\alias{flooding}
\title{Calculating flooding in Santa Barbara based on the rain precipitation values in the yearly recorded rainfall in Santa Barbara County for 2019}
\usage{
flooding(probability, land_size = 41.99, maximum = 2.02)
}
\arguments{
\item{with}{maximum recorded daily rainfall in Santa Barbara County in 2019}
}
\value{
with the return value of calculated risk with the amount that Santa Barbara County faces of a 100 year flood occuring within one year in Santa Barbara County in 2019
}
\description{
Calculating flooding in Santa Barbara based on the rain precipitation values in the yearly recorded rainfall in Santa Barbara County for 2019
}
\author{
: Nathaniel Burola
}
