% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimum.R
\name{minimum}
\alias{minimum}
\title{Function to find the smallest recorded precipitation value in the yearly recorded rainfall for Santa Barbara County in 2019}
\usage{
minimum(rainnum)
}
\arguments{
\item{daily_rain}{is precipitation recorded in Santa Barbara County in terms of inches}
}
\value{
with the return value finding the minimum amount of precipitation recorded
}
\description{
Function to find the smallest recorded precipitation value in the yearly recorded rainfall for Santa Barbara County in 2019
}
\author{
: Nathaniel Burola
}
